Modify main page structure
