+++ client/src/pages/main/index.tsx
@@ -1,17 +1,12 @@
 import React, { FC, useState, useEffect } from 'react';
 import * as S from 'pages/main/styles';
-import UserList from 'system/UserList';
 import NavigationBar from 'components/NavigationBar';
-import List from 'system/List';
 import shortid from 'shortid';
 import faker from 'faker';
-import RoomCard from 'components/RoomCard';
-import UserCard from 'components/UserCard';
-import Hr from 'commons/Hr';
+import Chat from './Chat';
+import Friend from './Friend';
 
-interface Props{
-
-}
+interface Props {}
 
 const Main: FC<Props> = () => {
   const [tabSelector, setTabSelector] = useState({
@@ -227,67 +222,10 @@ const Main: FC<Props> = () => {
             addFriendTabOnClick={addFriendTabOnClick}
             />
         </S.NavigationBarWrapper>
-        <S.Wrapper>
-          {
-            tabSelector.friend
-              && (
-                <List>
-                  <UserCard userName={myProfile.userName} />
-                  <Hr/>
-                  친구 {friendList.length}
-                  {
-                    friendList.map(({
-                      id,
-                      statusMessage,
-                      userName,
-                    }) => {
-                      const onUserCardClick = () => {
-                        alert(userName);
-                      };
-                      return (
-                        <UserCard
-                          key={id}
-                          userName={userName}
-                          statusMessage={statusMessage}
-                          onClick={onUserCardClick}
-                        />
-                      );
-                    })
-                  }
-                </List>
-              )
-          }
-          {
-            tabSelector.chat
-            && (
-              <List>
-                {
-                  chatList.map(({
-                    userList,
-                    lastMessage,
-                    lastModified,
-                    numOfNewMessage,
-                  }) => {
-                    const onRoomCardClick = () => {
-                      alert(userList.toString());
-                    };
-                    return (
-                      <RoomCard
-                        key={shortid.generate()}
-                        userList={userList}
-                        lastMessage={lastMessage}
-                        lastModified={lastModified}
-                        numOfNewMessage={numOfNewMessage}
-                        onClick={onRoomCardClick}
-                      />
-                    );
-                  })
-
-                }
-              </List>
-            )
-          }
-        </S.Wrapper>
+          <S.Wrapper>
+            { tabSelector.friend && <Friend myProfile={myProfile} friendList={friendList} /> }
+            { tabSelector.chat && <Chat chatList={chatList} /> }
+          </S.Wrapper>
       </S.Left>
     </S.Container>
   );
